Distributed Clock 'Cycle Shift' parameter should be a signed value
Enabling D.C. on an EtherCAT slave requires specifying a 'Cycle Time'
and 'Cycle Shift' parameter, where the shift will move the Sync pulse
either forward or backward relative to the 'Cycle Time'.
S.O.E.M. only allowed shifts in one direction because the 'CyclShift'
parameter was an unsigned integer -- to allow shifting the Sync in
either direction, it must be signed.