Move commonly used programs to utilities folder
The slaveinfo, simple_test and eepromtool programs are commonly used
for debugging and provisioning. This patch moves them to an
OS-independent subdirectory and builds them as part of the soem
library.

The patch also introduces the OSAL_MAIN define to allow platforms to
provide their own implementation of main.

Finally it adds the programs as shell commands for rt-kernel.