Return the "file not found" error for ecx_FOEread() through wkc
The "file not found" error is the most significant error that can be returned from the above function as it can affect and guide the work flow, so it makes sense to have it returned in some way. Since the same error has already been processed and returned in the ecx_FOEwrite() function, there is no reason not to process it and return it in the ecx_FOEread() function in the exact same way.